rename:logging_level_helper to log_level_helper
